Add files via upload
This is the first version of my program. There is no GUI, only console input and output, only hardcore!). But the Rome wasn't built in one day, gentlemen.
The data baze of clients of bank generates every time when you start program. This bank's traffic of clients is very dinamic!